% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello World}
\usage{
hello_world(name)
}
\arguments{
\item{name}{name to say hello to}
}
\value{
character string saying hello
}
\description{
Hello World
}
\section{Last Edited By}{
Ellis Hughes
}

\section{Last Edit Date}{
2021-03-12
}

